PythonProject/test01/requirements.txt
PyQt5~=5.15.11
opencv-python~=4.11.0.86
numpy~=2.2.4
PyGetWindow~=0.0.9
pillow~=11.1.0